+++ test/blackbox-tests/test-cases/with-stdin-from/run.t
@@ -1,0 +1,3 @@
+  $ dune runtest --display short
+           cat alias runtest
+  Hello!