+++ CHANGES.md
@@ -59,11 +59,14 @@
 - Add a new config option `sandboxing_preference`, the cli argument `--sandbox`,
   and the dep spec `sandbox` in dune language. These let the user control the level of
   sandboxing done by dune per rule and globally. The rule specification takes precedence.
-  The global configuration merely specifies the default. 
+  The global configuration merely specifies the default.
   (#2213, @aalekseyev, @jdimino)
 
 - Remove support for old style subsystems. Dune will now emit a warning to
   reinstall the library with the old style subsystem. (#2480, @rgrinberg)
+
+- Add action (with-stdin-from <file> <action>) to redirect input from <file>
+  when performing <action>. (#2487, @nojb)
 
 1.11.0 (23/07/2019)
 -------------------